Solution for day 3, 2019
A fun one because I use shapely all the time and it  hass really good
support for geometry intersection operations and the project method
easily allows you to find the distance a point is along a line.

Test driven development is going well, though today I separated tests
into their own files to make it quick and easy to run tests in
emacs. C-c C-t and elpy for the win!