Day 4 solution
Fun when you realise the numbers are misleading. The sort order for
text digits is the same as the numeric sort order. Don't have to worry
about type conversions here, just leave everything as strings.

Again, unit tests help for fixing dumb mistakes - in this case
forgettnig about the puzzle inputs of min and max values!